[REFACTORING] First round of refactoring for MVP prep. Decreasing size of main function, introduced QbicDataFinder class
